refactor(AddItemPage): Separate ItemForm component into another module
