+++ src/components/Market/ItemForm/ItemForm.module.css
@@ -25,163 +25,3 @@
   border-radius: 8px;
   cursor: pointer;
 }
-
-.submit-btn:disabled {
-  background-color: var(--cool-gray-400);
-}
-
-/*==========sub-title==========*/
-.sub-title {
-  color: var(--cool-gray-800);
-  font-size: 18px;
-  font-weight: 700;
-  line-height: 21px;
-  margin: 24px 0 12px;
-}
-
-.imgfile-submit > .sub-title {
-  margin: 16px 0 12px;
-}
-
-/*==========file-input==========*/
-.imgfile-container {
-  display: flex;
-  gap: 24px;
-}
-
-.file-input-label,
-.imgfile-preview {
-  width: 282px;
-  height: 282px;
-  border-radius: 12px;
-}
-
-.file-input-label {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--cool-gray-100);
-  cursor: pointer;
-}
-
-.register-img-icon {
-  display: block;
-  width: 74px;
-  height: 84px;
-}
-
-.file-input {
-  display: none;
-}
-
-.imgfile-preview {
-  display: none;
-  position: relative;
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-
-.imgfile-delete-btn {
-  display: block;
-  position: absolute;
-  top: 8px;
-  right: 8px;
-  width: 22px;
-  height: 22px;
-  background-size: 22px;
-  background-position: center;
-  background-repeat: no-repeat;
-  cursor: pointer;
-}
-
-.imgfile-delete-btn.active {
-  background-image: url("../../../assets/images/ic_img_delete_btn_active.png");
-}
-
-.imgfile-delete-btn.inactive {
-  background-image: url("../../../assets/images/ic_img_delete_btn_inactive.png");
-}
-
-/*==========text-input==========*/
-.name-input,
-.info-input,
-.price-input,
-.tag-input {
-  width: 100%;
-  color: var(--cool-gray-400);
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  background-color: var(--cool-gray-100);
-  border-radius: 12px;
-  padding: 16px 24px;
-}
-
-.name-input,
-.price-input,
-.tag-input {
-  height: 56px;
-}
-
-.info-input {
-  font-family: Pretendard, sans-serif;
-  height: 200px;
-  border: none;
-}
-
-/*==========tag-list==========*/
-.tags-container {
-  display: flex;
-  gap: 12px;
-  margin-top: 12px;
-}
-
-.tag {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  background-color: var(--cool-gray-50);
-  border-radius: 999px;
-  height: 48px;
-  padding: 12px 12px 12px 16px;
-}
-
-.tag-name {
-  color: var(--cool-gray-800);
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-}
-
-.tag-delete-btn {
-  display: block;
-  width: 22px;
-  height: 22px;
-  background-image: url("../../../assets/images/ic_img_delete_btn_inactive.png");
-  background-size: 22px;
-  background-position: center;
-  background-repeat: no-repeat;
-  cursor: pointer;
-}
-
-/*==========Tablet==========*/
-@media screen and (max-width: 1199px) {
-  .imgfile-container {
-    gap: 16px;
-  }
-
-  .file-input-label,
-  .imgfile-preview {
-    width: 162px;
-    height: 162px;
-    border-radius: 12px;
-  }
-}
-
-/*==========Mobile==========*/
-@media screen and (max-width: 767px) {
-  .imgfile-container {
-    gap: 8px;
-  }
-}
