+++ src/components/Market/ItemForm/ItemInfoInput.module.css
@@ -1,0 +1,70 @@
+/*==========sub-title==========*/
+.sub-title {
+  color: var(--cool-gray-800);
+  font-size: 18px;
+  font-weight: 700;
+  line-height: 21px;
+  margin: 24px 0 12px;
+}
+
+/*==========text-input==========*/
+.name-input,
+.info-input,
+.price-input,
+.tag-input {
+  width: 100%;
+  color: var(--cool-gray-400);
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+  background-color: var(--cool-gray-100);
+  border-radius: 12px;
+  padding: 16px 24px;
+}
+
+.name-input,
+.price-input,
+.tag-input {
+  height: 56px;
+}
+
+.info-input {
+  font-family: Pretendard, sans-serif;
+  height: 200px;
+  border: none;
+}
+
+/*==========tag-list==========*/
+.tags-container {
+  display: flex;
+  gap: 12px;
+  margin-top: 12px;
+}
+
+.tag {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  background-color: var(--cool-gray-50);
+  border-radius: 999px;
+  height: 48px;
+  padding: 12px 12px 12px 16px;
+}
+
+.tag-name {
+  color: var(--cool-gray-800);
+  font-size: 16px;
+  font-weight: 400;
+  line-height: 24px;
+}
+
+.tag-delete-btn {
+  display: block;
+  width: 22px;
+  height: 22px;
+  background-image: url("../../../assets/images/ic_img_delete_btn_inactive.png");
+  background-size: 22px;
+  background-position: center;
+  background-repeat: no-repeat;
+  cursor: pointer;
+}
